Working on cli
