+++ cli/old/main.py
@@ -1,0 +1,12 @@
+
+from color import *
+from cli.old.gameaccess import GameAccess
+from cli.old.user import userCli
+from cli.initGame 
+
+class cliManager
+
+
+# if __name__ == "__main__":
+#     user = userCli()
+#     game = GameAccess()
